+++ docs/plugins/multi-sidebar/components/ProjectSwitcher/ProjectSwitcher.astro
@@ -1,8 +1,8 @@
 ---
 import pluginConfig from 'virtual:starlight-multi-sidebar/config';
-import { getCurrentPathComponents } from '../path.ts';
-import { ProjectSwitcherReact } from './ProjectSwitcher.tsx';
-import { getCurrentHeader } from '../config.ts';
+import { getCurrentPathComponents } from '../../path.ts';
+import { ProjectSwitcherReact } from '../ProjectSwitcher/ProjectSwitcher.tsx';
+import { getCurrentHeader } from '../../config.ts';
 
 const { projectPath } = getCurrentPathComponents(Astro.url.pathname);
 const currentHeader = getCurrentHeader(pluginConfig, projectPath);
