+++ docs/plugins/multi-sidebar/components/ProjectSwitcher/ProjectSwitcher.css
