+++ docs/plugins/multi-sidebar/components/SiteTitle.astro
@@ -3,7 +3,7 @@
 
 import config from 'virtual:starlight/user-config';
 import { logos } from 'virtual:starlight/user-images';
-import ProjectSwitcher from './ProjectSwitcher.astro';
+import ProjectSwitcher from './ProjectSwitcher/ProjectSwitcher.astro';
 
 const { siteTitle, siteTitleHref } = Astro.locals.starlightRoute;
 ---
