+++ docs/plugins/multi-sidebar/components/VersionDropdown/VersionDropdown.css
